moving cli commands to directory suggested by mb
