+++ internal/pkg/cli/shell.go
